rename: entity 패키지 domain 패키지로 네이밍 변경
