+++ src/main/java/site/matzip/friendRequest/repository/FriendRequestRepository.java
@@ -2,7 +2,7 @@ package site.matzip.friendRequest.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import site.matzip.friendRequest.entity.FriendRequest;
+import site.matzip.friendRequest.domain.FriendRequest;
 import site.matzip.member.domain.Member;
 
 import java.util.List;
